refactor: Move the generator related code to its own submodule and the benchmark to its own.
