+++ ors-benchmark/src/main/java/org/heigit/ors/benchmark/AbstractLoadTest.java
@@ -4,7 +4,6 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import io.gatling.javaapi.core.PopulationBuilder;
 import io.gatling.javaapi.core.Simulation;
 import io.gatling.javaapi.http.HttpProtocolBuilder;
-import org.heigit.ors.config.Config;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
