+++ ors-benchmark/src/main/java/org/heigit/ors/benchmark/exceptions/CommandLineParsingException.java
@@ -1,0 +1,7 @@
+package org.heigit.ors.benchmark.exceptions;
+
+public class CommandLineParsingException extends RuntimeException {
+    public CommandLineParsingException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
