+++ ors-benchmark/src/main/java/org/heigit/ors/benchmark/exceptions/RequestBodyCreationException.java
@@ -1,0 +1,7 @@
+package org.heigit.ors.benchmark.exceptions;
+
+public class RequestBodyCreationException extends RuntimeException {
+    public RequestBodyCreationException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
