--- ors-benchmark/src/main/java/org/heigit/ors/exceptions/CommandLineParsingException.java
@@ -1,7 +1,0 @@
-package org.heigit.ors.exceptions;
-
-public class CommandLineParsingException extends RuntimeException {
-    public CommandLineParsingException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
