--- ors-benchmark/src/main/java/org/heigit/ors/exceptions/RequestBodyCreationException.java
@@ -1,7 +1,0 @@
-package org.heigit.ors.exceptions;
-
-public class RequestBodyCreationException extends RuntimeException {
-    public RequestBodyCreationException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
