+++ ors-benchmark/src/test/java/org/heigit/ors/util/ProgressBarLoggerTest.java
@@ -1,15 +1,12 @@
 package org.heigit.ors.util;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertSame;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import ch.qos.logback.classic.spi.ILoggingEvent;
+import ch.qos.logback.core.ConsoleAppender;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 
-import ch.qos.logback.classic.spi.ILoggingEvent;
-import ch.qos.logback.core.ConsoleAppender;
+import static org.junit.jupiter.api.Assertions.*;
 
 class ProgressBarLoggerTest {
     private Logger logger;
